Move all the things. (Do proper Python packaging.)
* Only use 'bikeshed' instead of 'lib' in the top-level Python namespace
* Use explicit relative imports (from .foo…)
* Move spec-data and include into the bikeshed package
* Add MANIFEST.in and fix setup.py to have data files installed
