+++ .gitignore
@@ -1,2 +1,3 @@
 *.egg-info
 *.pyc
+/dist
