+++ MANIFEST.in
@@ -1,0 +1,7 @@
+include README.md
+include Bikeshed.tmLanguage
+include docs/*
+include bikeshed/include/*
+include bikeshed/spec-data/*
+include bikeshed/widlparser/README.md
+
