+++ bikeshed/fuckunicode.py
