+++ bikeshed/include/copyright-personal.include
