+++ bikeshed/include/copyright.include
