+++ bikeshed/include/logo.include
