+++ bikeshed/widlparser/widlparser/constructs.py
