+++ bikeshed/widlparser/widlparser/tokenizer.py
