+++ setup.py
@@ -1,9 +1,10 @@
-from setuptools import setup
+from setuptools import setup, find_packages
 
 setup(
-    name='css-bikeshed',
+    name='Bikeshed',
     author='Tab Atkins Jr.',
-    py_modules=['preprocess'],
+    packages=find_packages(),
+    include_package_data=True,
     install_requires=['lxml', 'html5lib', 'cssselect'],
-    entry_points={'console_scripts': ['css-bikeshed = bikeshed:main']},
+    entry_points={'console_scripts': ['bikeshed = bikeshed:main']},
 )
